\name{58239904-package}
\alias{58239904-package}
\alias{58239904}
\docType{package}
\title{\packageTitle{58239904}}
\description{\packageDescription{58239904}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{58239904}
  \packageIndices{58239904}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{58239904}
  
  Maintainer: \packageMaintainer{58239904}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
